test_spack: Fix argument types in testcase
uint8_t obviously doesn't match the fmtstring for uint64_t.
Since it gets promoted to int, it still worked for uint{8,16,32}_t,
but it only worked for uint64_t by chance.

Should fix MSYS build error that appeared with a gcc update:
https://github.com/mmitch/gbsplay/actions/runs/6474003670/job/17577957685

+++ util.c
@@ -131,9 +131,9 @@ test void test_spack()
 	                   0x31, 1, 0, 2, 0, 0, 0, 3, 0, 0, 0, 0, 0, 0, 0, 4 };
 	uint8_t got[sizeof(want)] = { 0 };
 	uint8_t b = 1;
-	uint8_t w = 2;
-	uint8_t d = 3;
-	uint8_t q = 4;
+	uint16_t w = 2;
+	uint32_t d = 3;
+	uint64_t q = 4;
 
 	int l = spack(got, "{ }<bwdq{1}>bwdq", b, w, d, q, b, w, d, q);
